feat: refactor AttachmentController and remove Attachable concern for improved clarity and functionality 🎉✨
- Added before_action callbacks for ensure_enabled, attachment, context_crumbs, and current_page to streamline attachment handling.
- Implemented context_crumbs method to build breadcrumb navigation for attachments based on their parent objects (WorkflowExecution and SamplesWorkflowExecution) 🥖🔗.
- Created attachment_crumb method to generate breadcrumb items for the current attachment, enhancing navigation experience 📂.
- Removed the Attachable concern to simplify the controller structure and improve maintainability 🗑️.
